requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Migrate
pandas
scikit-learn
python-dotenv